Fixed the problem of usong  Double class in place of Decimal class in getElementAsDecimal method AXISCPP-675

+++ src/soap/SoapDeSerializer.cpp
@@ -2076,9 +2076,9 @@ xsd__decimal *
     SoapDeSerializer::getElementAsDecimal (const AxisChar * pName,
                                            const AxisChar * pNamespace)
 {
-    Double simpleType;
+    Decimal simpleType;
     getElement(pName, pNamespace, &simpleType);
-    return simpleType.getDouble();
+    return simpleType.getDecimal();
 }
 
 xsd__string
